+++ app/src/main/java/com/runwise/supply/firstpage/OrderModifyActivity.java
@@ -491,7 +491,7 @@ public class OrderModifyActivity extends NetWorkActivity implements OneKeyAdapte
             if (backList != null) {
                 for (AddedProduct pro : backList) {
                     Integer proId = Integer.valueOf(pro.getProductId());
-                    Integer count = pro.getCount();
+                    Integer count = (int)pro.getCount();
                     DefaultPBean bean = new DefaultPBean();
                     bean.setProductID(proId);
                     newList.add(bean);
