+++ app/src/main/java/com/runwise/supply/orderpage/entity/CreateCallInListRequest.java
@@ -19,13 +19,13 @@ public class CreateCallInListRequest {
     List<Product> products;
 
     public static class Product{
-        private int qty;
+        private double qty;
         private int productID;
-        public int getQty() {
+        public double getQty() {
             return qty;
         }
 
-        public void setQty(int qty) {
+        public void setQty(double qty) {
             this.qty = qty;
         }
